FIX: Fix a bug in levenshtein search
Depth tracking is necessary for levenshtein-search to produces as
little alignment matches as possible. The depth tracking in the
alignment search algorithm was broken for parital matches. This commit
fixes this deficiency.

The minimal depth in a search in a suffix tree is the depth we have to
at least reach in searching the suffix tree to return a result. It has
been set to the minimal length of an alignment result.

Generally it holds, that the higher the minimal depth, the faster search
runs.