FIX: Bug where slice index could be out of bounds for small token sizes
Searches with small minimal match lengths could lead to the
ma-search-service crashing because of the semantic analyzer creating a
vector slice that was out of bounds. The cause of this was, that the
java index creation worked with tokens from non-normalized language text
while the further processing steps all depended on the token vectors
originating from normalized text.