FEATURE: Now throw an exception and log properly on semantic problem
There is a persistent problem with semantic analysis where a vector
slice can be taken that exceeds the array bounds. Since I haven't yet
developed a fix as it's a bit complicated, the code now throws an
exception using MonadThrow, so that the backend-service can catch it
instead of just crashing.