REFACTOR: Make result set and index generic over annotation type
Up to now, search results were grouped by answers and in turn by an
answer-fragment.  This is actually not necessary. Instead, I introduced a
generic annotation type with which both index-tree-leaves and search results
could be tagged/grouped by.

This is relevant to allow easier testability down the line because quickly
trying stuff in the interpreter was not really possible before, which really
bogged down the development process.