FEATURE: Make python parser less picky with indentation
Up to now the python parser demanded, that code be indented by a consistent
number of spaces. It turns out that this breaks parsing for a large number of
code fragments in stackoverflow. Therefore this restriction is removed. This
will lead to the parser accepting a lot of potentially nonsensical code, but who
cares, it's stackoverflow.